requirements.txt
streamlit>=1.28.0
pytesseract>=0.3.10
opencv-python>=4.8.0
pillow>=10.0.0
openpyxl>=3.1.0
pyautogui>=0.9.54
pygetwindow>=0.0.9
numpy>=1.24.0
